pyproject.toml
[tool.poetry]
name = "evaluator"
version = "0.1.0"
description = ""
authors = ["Luis Vasquez <luis.vergara@teck.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pyyaml = "^6.0"
pandas = "^2.0.1"


[tool.poetry.group.pipeline.dependencies]
flake8 = "^6.0.0"
pylint = "^2.17.4"
pytest = "^7.3.1"
pytest-cov = "^4.0.0"
mypy = "^1.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
